ActuaryUtilities.jl

ActuaryUtilities.InterestCurve

ActuaryUtilities.breakeven

ActuaryUtilities.breakeven

ActuaryUtilities.discount_rate

ActuaryUtilities.duration

ActuaryUtilities.duration

ActuaryUtilities.duration

ActuaryUtilities.duration

ActuaryUtilities.interest_rate

ActuaryUtilities.internal_rate_of_return

ActuaryUtilities.internal_rate_of_return

ActuaryUtilities.irr

ActuaryUtilities.present_value

ActuaryUtilities.present_value

ActuaryUtilities.pv

ActuaryUtilities.years_between

ActuaryUtilities.InterestCurve — Method

InterestCurve(rates,times,interpolation_method)

Creates an InterestCurve object:

rates are annual effective forward interest rates

times are the end of the period for which the

interpolation_method is either option below (with StepwiseInterp() as the default)

StepwiseInterp() which is an ActuaryUtilities.jl provided type which will use the given rate for the period up to the given time.

For example, with rates= [0.05,0.1] and times=[1,2] then 0.05 will be used for the period (0,1] and 0.1 will be used for the period (1,2].

LinearInterp(), which linearly interpolates between times and is flat outside the boundaries of times.

ActuaryUtilities.breakeven — Method

breakeven(cashflows::Vector,accumulation_rate::Real)

Calculate the time when the accumulated cashflows breakeven. Assumes that :

cashflows evenly spaced with the first one occuring at time zero

cashflows occur at the end of the period

that the accumulation rate correponds to the periodicity of the cashflows.

Returns nothing if cashflow stream never breaks even.

julia> breakeven([-10,1,2,3,4,8],0.10)
5

julia> breakeven([-10,15,2,3,4,8],0.10)
1

julia> breakeven([-10,-15,2,3,4,8],0.10) # returns the `nothing` value

ActuaryUtilities.breakeven — Method

breakeven(cashflows::Vector,timepoints::Vector, accumulation_rate)

Calculate the time when the accumulated cashflows breakeven. Assumes that:

cashflows occur at the timepoint indicated at the corresponding timepoints position

cashflows occur at the end of the period

that the accumulation rate corresponds to the periodicity of the cashflows.

If given a vector of

interest rates, the first rate is effectively never used, as it's treated as the accumulation rate between time zero and the first cashflow.

Returns nothing if cashflow stream never breaks even.

julia> times = [0,1,2,3,4,5];

julia> breakeven([-10,1,2,3,4,8],times,0.10)
5

julia> breakeven([-10,15,2,3,4,8],times,0.10)
1

julia> breakeven([-10,-15,2,3,4,8],times,0.10) # returns the `nothing` value

ActuaryUtilities.discount_rate — Method

discount_rate(interest,time)

Return the discount rate at time.

interest can be: - a InterestCurve - a vector wrapped in an interest curve - a scalar rate

Internally, if not a scalar argument, this method will use the interpolated interest rate to use an integral approxmation to the accumulated force of interest. This generalizes well. For more performance on repeated calls, use an InterestCurve instead of a vector.

Examples

InterestCurve

julia> ic = InterestCurve([0.01,0.05,0.05,0.1],[1,2,3,4])
julia> ic = InterestCurve([0.01,0.05,0.05,0.1])  # this is equivalent to the line above

julia> discount_rate(ic,1)
0.9900990099009901

julia> discount_rate.(ic,0:4) # function is broadcastable
5-element Array{Float64,1}:
 1.0
 0.9900990099009901
 0.9613215887833819
 0.9155443705701612
 0.8517459660161829

Scalar

julia> discount_rate(0.05,1)
0.9523809523809523

julia> discount_rate.(0.05,1:3) # function can be broadcasted
3-element Array{Float64,1}:
 0.9523809523809523
 0.9070294784580498
 0.863837598531476

ActuaryUtilities.duration — Method

duration(::DV01,int,cfs,times)

Calculates the DV01, with a fixed int (e.g. 0.05). int should be an annual effective interest rate, not a bond effective yield.

Not yet generalized for InterestCurve arguments

ActuaryUtilities.duration — Method

duration(d1::Date, d2::Date)

Compute the duration given two dates, which is the number of years since the first date. The interval [0,1) is defined as having duration 1. Can return negative durations if second argument is before the first.

julia> issue_date  = Date(2018,9,30);

julia> duration(issue_date , Date(2019,9,30) ) 
2
julia> duration(issue_date , issue_date) 
1
julia> duration(issue_date , Date(2018,10,1) ) 
1
julia> duration(issue_date , Date(2019,10,1) ) 
2
julia> duration(issue_date , Date(2018,6,30) ) 
0
julia> duration(Date(2018,9,30),Date(2017,6,30)) 
-1

ActuaryUtilities.duration — Method

duration(::Macaulay,int,cfs,times)

Calculates the Macaulay duration, with a fixed int (e.g. 0.05). int should be an annual effective interest rate, not a bond effective yield.

Not yet generalized for InterestCurve arguments.

ActuaryUtilities.duration — Method

duration(::Modified,int,cfs,times)

Calculates the Modified duration, with a fixed int (e.g. 0.05). int should be an annual effective interest rate, not a bond effective yield.

Not yet generalized for InterestCurve arguments

ActuaryUtilities.interest_rate — Method

interest_rate(interest,time)

Return the interest rate at time.

interest can be: - an InterestCurve - a vector wrapped in an interest curve - a scalar rate

Examples

InterestCurve

julia> ic = InterestCurve([0.01,0.05,0.05,0.1],[1,2,3,4])
julia> ic = InterestCurve([0.01,0.05,0.05,0.1])  # this is equivalent to the line above

julia> interest_rate(ic,1)
0.01

julia> interest_rate.(ic,0:4) # function is broadcastable
5-element Array{Float64,1}:
 0.01
 0.01
 0.05
 0.05
 0.1

Scalar

julia> interest_rate(0.05,1)
0.05

julia> interest_rate.(0.05,1:3) # function can be broadcasted
3-element Array{Float64,1}:
0.05
0.05
0.05

ActuaryUtilities.internal_rate_of_return — Method

internal_rate_of_return(cashflows::Vector, timepoints::Vector)

Calculate the internalrateof_return with given timepoints. First tries to find a positive rate in the interval [0.0,1.0]. If none is found, will extend search to [-1.0,1.0]. If still not found, will return nothing.

julia> internal_rate_of_return([-100,110],[0,1]) # e.g. cashflows at time 0 and 1
0.10000000001652906

ActuaryUtilities.internal_rate_of_return — Method

internal_rate_of_return(cashflows::vector)

Calculate the internalrateof_return of a series of equally spaced cashflows, assuming the first element occurs at time zero. First tries to find a positive rate in the interval [0.0,1.0]. If none is found, will extend search to [-1.0,1.0]. If still not found, will return nothing.

ActuaryUtilities.irr — Function

irr()

An alias for `internal_rate_of_return`.

ActuaryUtilities.present_value — Method

present_value(interest, cashflows::Vector)

Assumes that cashflows happen at times 1,2,...,n.

interest can be an InterestCurve, a single scalar, or a vector.

ActuaryUtilities.present_value — Method

present_value(interest, cashflows::Vector, timepoints)

Discount the cashflows vector at the given interest_interestrate, with the cashflows occurring at the times specified in timepoints.

The interest can be an InterestCurve, a single scalar, or a vector wrapped in an InterestCurve.

Examples

julia> present_value(0.1, [10,20],[0,1])
28.18181818181818
julia> present_value(InterestVector([0.1,0.2]), [10,20],[0,1])

Example on how to use real dates using the DayCounts.jl package

using DayCounts 
dates = Date(2012,12,31):Year(1):Date(2013,12,31)
times = map(d -> yearfrac(dates[1], d, DayCounts.Actual365Fixed()),dates) # [0.0,1.0]
present_value(0.1, [10,20],times)

# output
28.18181818181818

ActuaryUtilities.pv — Function

pv()

An alias for `present_value`.

ActuaryUtilities.years_between — Function

Years_Between(d1::Date, d2::Date)

Compute the number of integer years between two dates, with the first date typically before the second. Will return negative number if first date is after the second. Use third argument to indicate if calendar anniversary should count as a full year.

Examples

julia> d1 = Date(2018,09,30);

julia> d2 = Date(2019,09,30);

julia> d3 = Date(2019,10,01);

julia> years_between(d1,d3) 
1
julia> years_between(d1,d2,false) # same month/day but `false` overlap
0 
julia> years_between(d1,d2) # same month/day but `true` overlap
1 
julia> years_between(d1,d2) # using default `true` overlap
1